+++ hack/README.md
@@ -1,0 +1,6 @@
+# NOTICE: 
+
+## **Tools in this directory are provided as is with no SLA, and not for production use.** They are purely experiments or hackathon projects that have no home.
+
+### Thanks,
+### Azure Container Networking Dev Team
