+++ hack/acncli/.dockerignore
