+++ hack/acncli/installer/utils.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"path/filepath"
 
-	c "github.com/Azure/azure-container-networking/acncli/api"
+	c "github.com/Azure/azure-container-networking/hack/acncli/api"
 )
 
 func SetOrUseDefault(setValue, defaultValue string) string {
