+++ hack/acncli/main.go
@@ -1,12 +1,12 @@
 package main
 
 import (
-	"github.com/Azure/azure-container-networking/acncli/cmd"
+	"github.com/Azure/azure-container-networking/hack/acncli/cmd"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
 
-	c "github.com/Azure/azure-container-networking/acncli/api"
+	c "github.com/Azure/azure-container-networking/hack/acncli/api"
 )
 
 var (
